@@ -1,6 +1,6 @@
 {
     "name": "zfr/zfr-cors",
-    "description": "Zend Framework 2 module that let you deal with CORS requests",
+    "description": "Zend Framework module that let you deal with CORS requests",
     "type": "library",
     "license": "MIT",
     "keywords": [
@@ -28,7 +28,7 @@
         }
     },
     "require": {
-        "php": "^5.6 || ^7.0",
+        "php": "^7.1",
         "zendframework/zend-eventmanager": "^2.6.3 || ^3.0.1",
         "zendframework/zend-http": "^2.5.5",
         "zendframework/zend-mvc": "^2.7.9 || ^3.0.2",
@@ -40,9 +40,9 @@
         "zendframework/zend-modulemanager": "^2.7.2",
         "zendframework/zend-serializer": "^2.8",
         "zendframework/zend-view": "^2.8.1",
-        "phpunit/phpunit": "^4.8 || ^5.0",
-        "squizlabs/php_codesniffer": "^2.6.2",
-        "satooshi/php-coveralls": "^1.0.1"
+        "phpunit/phpunit": "^7.5.9 || ^8.2.2",
+        "squizlabs/php_codesniffer": "^3.4.2",
+        "satooshi/php-coveralls": "^2.1"
     },
     "autoload": {
         "psr-0": {
